Move initializer utils to a dedicated module
